Change calls to std::make_tuples into calls to std::tuple's constructor
Changing the calls from std::make_tuple to std::tuple's constructor is
required to work around an MSVC bug.

When explicitly specifying the type parameters to a template function
that takes forwarding references as argument, MSVC fails to properly
perform array-to-pointer decay. In this case, we are explicitly
specifying
the first type template argument to be `const char*`. So, the type of
the
first argument to std::make_tuple becomes `const char*&&`. However,
when passing a string literal as the first argument to std::make_tuple,
MSVC fails to decay the type of the string literal from `char[]` to
`const char*`, causing a compilation error.

Changing std::make_tuple to std::tuple avoids the error because the
std::tuple constructor has overloads that do not take forwarding
references as arguments, causing array-to-pointer decay to happen
correctly.

Signed-off-by: Bohao(Aaron) Wang <aaronwang0407@gmail.com>